requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
celery>=5.3
redis>=5.0
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.2
psycopg2-binary>=2.9
Pillow>=10.0
